requirements.txt
# M0/M1 Card Terminal - Production Dependencies
# ==============================================

# Core Flask Framework
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0

# Security & Rate Limiting
flask-cors==4.0.0
flask-limiter==3.5.0

# Blockchain Integration
web3==6.9.0
tronpy==0.4.0

# Cryptography
cryptography==41.0.3
eth-account==0.9.0
eth-utils==2.2.0
eth-typing==3.4.0
eth-hash==0.5.2
eth-keys==0.4.0
eth-rlp==0.3.0
eth-abi==4.2.1

# Utilities
hexbytes==0.3.1
bitarray==2.8.1
protobuf==4.24.0
base58==2.1.1
ecdsa==0.18.0

# Production Server
gunicorn==21.2.0

# Environment Management
python-dotenv==1.0.0
